+++ src/library_browser.js
@@ -832,27 +832,27 @@ var LibraryBrowser = {
   emscripten_async_run_script__deps: ['emscripten_run_script', '$safeSetTimeout'],
   emscripten_async_run_script: function(script, millis) {
     // TODO: cache these to avoid generating garbage
-    safeSetTimeout(function() {
-      _emscripten_run_script(script);
-    }, millis);
+    safeSetTimeout(() => _emscripten_run_script(script), millis);
   },
 
   // TODO: currently not callable from a pthread, but immediately calls onerror() if not on main thread.
+  emscripten_async_load_script__deps: ['$UTF8ToString'],
   emscripten_async_load_script: function(url, onload, onerror) {
+    url = UTF8ToString(url);
     onload = {{{ makeDynCall('v', 'onload') }}};
     onerror = {{{ makeDynCall('v', 'onerror') }}};
 
 #if PTHREADS
     if (ENVIRONMENT_IS_PTHREAD) {
-      err('emscripten_async_load_script("' + UTF8ToString(url) + '") failed, emscripten_async_load_script is currently not available in pthreads!');
+      err('emscripten_async_load_script("' + url + '") failed, emscripten_async_load_script is currently not available in pthreads!');
       return onerror ? onerror() : undefined;
     }
 #endif
+    assert(runDependencies === 0, 'async_load_script must be run when no other dependencies are active');
+
     {{{ runtimeKeepalivePush() }}}
 
-    assert(runDependencies === 0, 'async_load_script must be run when no other dependencies are active');
-    var script = document.createElement('script');
-    script.onload = function script_onload() {
+    var loadDone = () => {
       {{{ runtimeKeepalivePop() }}}
       if (onload) {
         if (runDependencies > 0) {
@@ -861,12 +861,27 @@ var LibraryBrowser = {
           onload();
         }
       }
-    };
-    script.onerror = () => {
+    }
+
+    var loadError = () => {
       {{{ runtimeKeepalivePop() }}}
       if (onerror) onerror();
     };
-    script.src = UTF8ToString(url);
+
+#if ENVIRONMENT_MAY_BE_NODE && DYNAMIC_EXECUTION
+    if (ENVIRONMENT_IS_NODE) {
+      readAsync(url, (data) => {
+        eval(data);
+        loadDone();
+      }, loadError, false);
+      return;
+    }
+#endif
+
+    var script = document.createElement('script');
+    script.onload = loadDone;
+    script.onerror = loadError;
+    script.src = url;
     document.body.appendChild(script);
   },